requirements.txt
torch>=1.13.0
torchvision>=0.14.0
flask>=2.0.1
numpy>=1.21.2
matplotlib>=3.4.3
scikit-learn>=1.0.1
pillow>=8.3.2
grad-cam>=1.3.7
kaggle
tqdm
seaborn